refactor: Add skeletal method for marking attendance
Notes:

* Need to progressively add requirements of ID class and the required command for marking attendance
