+++ src/main/java/seedu/address/model/person/Person.java
@@ -2,8 +2,10 @@ package seedu.address.model.person;
 
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
 
+import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Objects;
 import java.util.Set;
 
@@ -24,6 +26,7 @@ public class Person {
     // Data fields
     private final Address address;
     private final Set<Tag> tags = new HashSet<>();
+    private final List<Attendance> attendanceRecords = new ArrayList<>();
 
     /**
      * Every field must be present and not null.
@@ -72,6 +75,10 @@ public class Person {
 
         return otherPerson != null
                 && otherPerson.getName().equals(getName());
+    }
+
+    public void addAttendance(Attendance attendance) {
+        attendanceRecords.add(attendance);
     }
 
     /**
